chore: Remove unused TTL files from triplestore
This commit removes TTL files from the triplestore that do not have a corresponding graph in the triplestore. It adds a loop to iterate through the graphs in the triplestore and checks if there is a corresponding TTL file. If not, it drops the graph from the triplestore. This ensures that the triplestore only contains graphs that have associated TTL files.